moving specs_checkers.py logic into validators.py
